refactor(code): move Textual adapter into tui package
